requirements.txt
streamlit==1.31.0
anthropic>=0.25.0
python-dotenv==1.0.0
requests==2.31.0
facebook-sdk==3.1.0
python-linkedin-v2==0.9.1
instagrapi==2.0.0
tweepy==4.14.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.100.0
plotly==5.17.0
pandas==2.1.3
schedule==1.2.0
scikit-learn==1.3.2
numpy==1.24.3
kaggle==1.5.16